Moved all source files to respect javafx gradle working tree's
This commit starts feature #1 with a minimalist but working gradle build file's
