+++ build.gradle
@@ -5,7 +5,6 @@ buildscript {
     repositories {
         mavenLocal()
         mavenCentral()
-
     }
 }
 apply plugin: 'java'
@@ -17,11 +16,19 @@ repositories {
 
 apply plugin: 'javafx-gradle-plugin'
 
+dependencies {
+    compile 'org.apache.logging.log4j:log4j-api:2.10.0'
+    compile 'org.apache.logging.log4j:log4j-core:2.10.0'
+    compile 'org.xerial:sqlite-jdbc:3.21.0.1'
+}
+
+buildScan {
+    licenseAgreementUrl = 'https://gradle.com/terms-of-service';
+    licenseAgree = 'no'
+}
 
 jfx {
-    // minimal requirement for jfxJar-task
     mainClass = 'it.bogny.jyugiohdb.MainApp'
-    additionalAppResources = 'src'
-    // minimal requirement for jfxNative-task
+    jfxMainAppJarName = 'jyugiohdb.jar'
     vendor = 'BoGnY'
 }
