--- src/it/bogny/jyugiohdb/database/package-info.java
@@ -1,12 +1,0 @@
-/**
- *
- * This package contains all classes that manage connections with the SQLite
- * database
- *
- * @author BoGnY
- * 
- * @since 0.1
- * 
- */
-
-package it.bogny.jyugiohdb.database;
