--- src/it/bogny/jyugiohdb/model/package-info.java
@@ -1,11 +1,0 @@
-/**
- *
- * This package contains all the model classes
- *
- * @author BoGnY
- * 
- * @since 0.1
- * 
- */
-
-package it.bogny.jyugiohdb.model;
