--- src/it/bogny/jyugiohdb/util/package-info.java
@@ -1,11 +1,0 @@
-/**
- *
- * This package contains all the utility classes for the application
- *
- * @author BoGnY
- * 
- * @since 0.1
- * 
- */
-
-package it.bogny.jyugiohdb.util;
