--- src/it/bogny/jyugiohdb/view/package-info.java
@@ -1,11 +1,0 @@
-/**
- *
- * This package contains all the view classes and its controllers
- *
- * @author BoGnY
- * 
- * @since 0.1
- * 
- */
-
-package it.bogny.jyugiohdb.view;
