+++ src/main/java/it/bogny/jyugiohdb/database/package-info.java
@@ -1,0 +1,12 @@
+/**
+ *
+ * This package contains all classes that manage connections with the SQLite
+ * database
+ *
+ * @author BoGnY
+ * 
+ * @since 0.1
+ * 
+ */
+
+package it.bogny.jyugiohdb.database;
