+++ src/main/java/it/bogny/jyugiohdb/package-info.java
@@ -1,0 +1,11 @@
+/**
+ *
+ * This package contains core of application
+ *
+ * @author BoGnY
+ *
+ * @version 0.1.0.18
+ * 
+ */
+
+package it.bogny.jyugiohdb;
