+++ src/main/resources/it/bogny/jyugiohdb/css/application.css
