+++ src/main/resources/it/bogny/jyugiohdb/images/app_icon.png
