+++ src/main/resources/it/bogny/jyugiohdb/view/CardOverview.fxml
