+++ src/main/resources/it/bogny/jyugiohdb/view/MainLayout.fxml
